chore: [#184606287] prefactor - move layout files into layout folder
