+++ web/src/components/ModalTwoButton.tsx
@@ -1,7 +1,7 @@
 import { ModalZeroButton } from "@/components/ModalZeroButton";
 import { PrimaryButton } from "@/components/njwds-extended/PrimaryButton";
 import { SecondaryButton } from "@/components/njwds-extended/SecondaryButton";
-import { ReverseOrderInMobile } from "@/components/ReverseOrderInMobile";
+import { ReverseOrderInMobile } from "@/components/njwds-layout/ReverseOrderInMobile";
 import { MediaQueries } from "@/lib/PageSizes";
 import { Breakpoint, useMediaQuery } from "@mui/material";
 import { ReactElement, ReactNode } from "react";
