+++ web/src/pages/404.tsx
@@ -1,7 +1,7 @@
 import { Content } from "@/components/Content";
 import { UnStyledButton } from "@/components/njwds-extended/UnStyledButton";
+import { PageSkeleton } from "@/components/njwds-layout/PageSkeleton";
 import { SingleColumnContainer } from "@/components/njwds/SingleColumnContainer";
-import { PageSkeleton } from "@/components/PageSkeleton";
 import { useConfig } from "@/lib/data-hooks/useConfig";
 import { ReactElement } from "react";
