+++ web/src/pages/funding/[fundingUrlSlug].tsx
@@ -1,7 +1,7 @@
 import { Content } from "@/components/Content";
 import { HorizontalLine } from "@/components/HorizontalLine";
 import { NavBar } from "@/components/navbar/NavBar";
-import { PageSkeleton } from "@/components/PageSkeleton";
+import { PageSkeleton } from "@/components/njwds-layout/PageSkeleton";
 import { TaskCTA } from "@/components/TaskCTA";
 import { TaskSidebarPageLayout } from "@/components/TaskSidebarPageLayout";
 import { useConfig } from "@/lib/data-hooks/useConfig";
